#50 refactor: actor-settings の型・定数を types.ts/constants.ts へ分離
- ActorSettings 系の型定義を types.ts、DEFAULT_ACTOR_SETTINGS を constants.ts に切り出す
- store.ts は実装のみ残し、型・定数はそこから import する形にする
- 他 store が実装 (create関数) ではなく型・定数のみに依存する構造にするための準備
