+++ src/prototypes/time-control/time-control-03/_stores/actor-settings/constants.ts
@@ -1,0 +1,8 @@
+import { ActorSettings } from './types'
+
+/** 操作設定未設定 actor のデフォルト値 */
+export const DEFAULT_ACTOR_SETTINGS: ActorSettings = {
+  fixedPathSteps: 1,
+  isFixedPathSteps: false,
+  progressMode: 'auto',
+}
