+++ src/prototypes/time-control/time-control-03/_stores/actor-settings/context.tsx
@@ -1,6 +1,6 @@
 import { createStoreContext } from '@/stores/utils/create-store-context'
 
-import { ActorSettingsState } from './store'
+import { ActorSettingsState } from './types'
 
 const { StoreContext, useStoreSelector } =
   createStoreContext<ActorSettingsState>('ActorSettings')
